third_party/blink/web_tests/fast/forms/text/placeholder-dom-property-expected.txt

Test for https://bugs.webkit.org/show_bug.cgi?id=21903

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS input.placeholder is ""
PASS input.placeholder is "p1"
PASS input.getAttribute('placeholder') is "p2"
PASS input.placeholder is "p2"
PASS successfullyParsed is true

TEST COMPLETE
